Move overload test back into single test class
