+++ tests/Client/JustGivingClientTest.php
@@ -2,6 +2,7 @@
 
 namespace Konsulting\JustGivingApiSdk\Tests\Client;
 
+use GuzzleHttp\Client;
 use GuzzleHttp\Psr7\Response;
 use Konsulting\JustGivingApiSdk\Exceptions\ClassNotFoundException;
 use Konsulting\JustGivingApiSdk\JustGivingClient;
@@ -81,6 +82,29 @@ class JustGivingClientTest extends TestCase
             'root_domain' => 'https://example.com',
             'api_version' => 3,
         ]);
+
+        $client->Account->retrieve();
+    }
+
+    /**
+     * We're using an overload mock, which will fail if the class has already been loaded for another test.
+     *
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     * @test
+     */
+    public function it_automatically_instantiates_a_psr_18_client()
+    {
+        // Overload the Guzzle client class to make sure the JustGiving client is instantiating it internally
+        $http = Mockery::mock('overload:' . Client::class);
+        $http->shouldReceive('send')
+            ->withArgs(function (RequestInterface $request) {
+                return $request->getUri()->__toString() === 'https://api.justgiving.com/v1/account';
+            })
+            ->once()
+            ->andReturn(new Response);
+
+        $client = new JustGivingClient($this->getAuthMock());
 
         $client->Account->retrieve();
     }
